refactor(backend-runtime): extract control_mode module, slim main.rs to 201 lines
- Extract run_control_mode + handle_control_connection into control_mode.rs
- Extract spawn_driver_sampler helper function
- Add control_mode module-level unit tests (2 new tests)
- Fix SharedHandles wiring in ControlSpawnKit
- Suppress dead_code warning on PlayerLoopShared fields
- main.rs: 201 lines (was 602) — under the 350-line acceptance target
- All 12 tests pass, cargo fmt clean, zero warnings
